pyproject.toml
[project]
name = "dspyui"
version = "0.1.0"
description = "A Gradio user interface for DSPy"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "gradio==6.2.0",
    "dspy-ai==3.0.4",
    "pandas==2.3.3",
    "openai==2.14.0",
    "anthropic==0.75.0",
    "groq==1.0.0",
    "scikit-learn==1.8.0",
    "google-generativeai==0.8.6",
    "numpy==2.4.0",
    "pydantic==2.12.5",
    "python-dotenv==1.2.1",
]

[project.scripts]
dspyui = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.hatch.build.targets.wheel]
packages = ["dspyui"]

